requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dashscope==1.25.12
fastapi==0.104.1
frozenlist==1.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
numpy==2.4.2
pandas==3.0.0
pillow==12.1.1
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
yarl==1.22.0
